Create common app to hold assets used by other apps. Include other minor changes.
