+++ common/tests/test_error_views.py
@@ -1,4 +1,4 @@
-"""Index view tests.
+"""Common test error views.
 
 Author: Jared Paubel
 Version: 0.1
@@ -6,14 +6,14 @@ Version: 0.1
 from django.core.exceptions import BadRequest, PermissionDenied
 from requests import HTTPError
 from django.test import TestCase, RequestFactory
-from index.views import (
+from common.views import (
     custom_page_not_found_view, custom_bad_request_view,
     custom_error_view, custom_permission_denied_view
 )
 
 
-class IndexViewTest(TestCase):
-    """Index view test class."""
+class TestErrorViews(TestCase):
+    """Test common views."""
 
     def setUp(self):
         self.factory = RequestFactory()
