+++ index/urls.py
@@ -5,9 +5,8 @@ Version: 0.1
 """
 from django.urls import path
 
-from index.views import index_view, search
+from index.views import index_view
 
 urlpatterns = [
     path('', index_view, name='home'),
-    path('search/', search, name='search'),
 ]
