+++ prokope/urls.py
@@ -8,14 +8,15 @@ from django.urls import path, include
 if not settings.DEBUG:
     admin.site.__class__ = OTPAdminSite
 
-handler400 = 'index.views.custom_bad_request_view'
-handler403 = 'index.views.custom_permission_denied_view'
-handler404 = 'index.views.custom_page_not_found_view'
-handler500 = 'index.views.custom_error_view'
+handler400 = 'common.views.custom_bad_request_view'
+handler403 = 'common.views.custom_permission_denied_view'
+handler404 = 'common.views.custom_page_not_found_view'
+handler500 = 'common.views.custom_error_view'
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('index.urls')),
+    path('', include('common.urls')),
     path('blog', include('blog.urls')),
     path('gallery', include('gallery.urls')),
     path('summernote/', include('django_summernote.urls')),
