+++ runserver.sh
@@ -74,7 +74,7 @@ else
 fi
 
 echo "🧪 Running tests..."
-coverage run --source=index --omit=*/migrations/* manage.py test index
+coverage run --source=common,index --omit=*/migrations/* manage.py test common index
 
 # Check if tests failed
 if [ $? -ne 0 ]; then
